change the currency_value field to float in the transactions table

+++ server/database/migrations/2021_08_17_121446_create_transactions_table.php
@@ -19,7 +19,7 @@ class CreateTransactionsTable extends Migration
             $table->unsignedInteger("currency_id")->nullable();
             $table->foreign("currency_id")->references("id")->on("currencies");
 
-            $table->integer("currency_value");
+            $table->float("currency_value");
 
             $table->boolean("type");
